Stop feature:ui reaching past the bus into processor internals
VariableAmountHandler used to call NotificationProcessor.setReminderEventStatus
directly. Route it through ReminderCommandBus.markReminderEvents instead:
persist the variable-amount edit via the repository, then mark the event TAKEN
via the bus (which re-hydrates from the DB and runs the same setReminderEventStatus
path internally).

ReminderEventCreator used to call ReminderNotificationProcessor.buildReminderEvent
on a companion object. Extract that helper to a top-level suspend function in
feature:reminders/ReminderEventFactory.kt. feature:ui now imports the free
function instead of the processor class. Update the in-module caller
(ReminderNotificationFactory) to use the same top-level function.

feature:ui no longer imports NotificationProcessor or ReminderNotificationProcessor.
The classes remain public for now because cross-module unit tests in app/src/test
still hold direct references; marking them internal is a follow-up that requires
moving those tests into feature:reminders.
