+++ feature/reminders/src/main/java/com/futsch1/medtimer/feature/reminders/ReminderNotificationProcessor.kt
@@ -6,21 +6,13 @@ import android.content.pm.PackageManager
 import android.os.Build
 import android.util.Log
 import com.futsch1.medtimer.core.common.LogTags
-import com.futsch1.medtimer.core.common.helpers.MedicineHelper
-import com.futsch1.medtimer.core.common.helpers.TimeHelper
 import com.futsch1.medtimer.core.datastore.PreferencesDataSource
-import com.futsch1.medtimer.core.domain.model.Medicine
-import com.futsch1.medtimer.core.domain.model.Reminder
 import com.futsch1.medtimer.core.domain.model.ReminderEvent
-import com.futsch1.medtimer.core.domain.model.ReminderType
 import com.futsch1.medtimer.core.domain.repository.ReminderEventRepository
-import com.futsch1.medtimer.core.ui.TimeFormatter
 import com.futsch1.medtimer.feature.reminders.notificationData.ReminderNotification
 import com.futsch1.medtimer.feature.reminders.notificationData.ReminderNotificationData
 import com.futsch1.medtimer.feature.reminders.notificationData.ReminderNotificationFactory
-import com.futsch1.medtimer.feature.reminders.scheduling.CyclesHelper
 import dagger.hilt.android.qualifiers.ApplicationContext
-import java.time.Instant
 import javax.inject.Inject
 
 class ReminderNotificationProcessor @Inject constructor(
@@ -101,80 +93,5 @@ class ReminderNotificationProcessor @Inject constructor(
 
     private fun canShowNotifications(): Boolean {
         return Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU || context.checkSelfPermission(Manifest.permission.POST_NOTIFICATIONS) == PackageManager.PERMISSION_GRANTED
-    }
-
-    companion object {
-        suspend fun buildReminderEvent(
-            remindedTimeStamp: Long,
-            medicine: Medicine,
-            reminder: Reminder,
-            reminderEventRepository: ReminderEventRepository,
-            timeFormatter: TimeFormatter
-        ): ReminderEvent {
-            val remindedInstant = Instant.ofEpochSecond(remindedTimeStamp)
-            val amount = when (reminder.reminderType) {
-                ReminderType.OUT_OF_STOCK -> {
-                    MedicineHelper.formatAmount(medicine.amount, medicine.unit)
-                }
-
-                ReminderType.EXPIRATION_DATE -> {
-                    timeFormatter.localDateToString(medicine.expirationDate)
-                }
-
-                else -> {
-                    reminder.amount
-                }
-            }
-
-            val lastIntervalReminderTimeInMinutes = if (reminder.isInterval) {
-                getLastReminderEventTimeInMinutes(
-                    reminderEventRepository,
-                    reminder.id,
-                    remindedInstant,
-                    reminder.reminderType == ReminderType.WINDOWED_INTERVAL
-                )
-            } else {
-                0
-            }
-
-            return ReminderEvent(
-                reminderEventId = 0,
-                reminderId = reminder.id,
-                medicineName = medicine.name + CyclesHelper.getCycleCountString(reminder),
-                amount = amount,
-                color = medicine.color,
-                useColor = medicine.useColor,
-                status = ReminderEvent.ReminderStatus.RAISED,
-                remindedTimestamp = remindedInstant,
-                processedTimestamp = Instant.EPOCH,
-                notificationId = 0,
-                iconId = medicine.iconId,
-                remainingRepeats = 0,
-                notes = "",
-                reminderType = reminder.reminderType,
-                stockHandled = false,
-                askForAmount = reminder.variableAmount,
-                tags = medicine.tags.map { it.name },
-                lastIntervalReminderTimeInMinutes = lastIntervalReminderTimeInMinutes
-            )
-        }
-
-        private suspend fun getLastReminderEventTimeInMinutes(
-            reminderEventRepository: ReminderEventRepository,
-            reminderId: Int,
-            remindedTimestamp: Instant,
-            isWindowedInterval: Boolean
-        ): Int {
-            val lastReminderEvent = reminderEventRepository.getLast(reminderId) ?: return 0
-            if (lastReminderEvent.status != ReminderEvent.ReminderStatus.TAKEN) return 0
-
-            if (isWindowedInterval &&
-                TimeHelper.isSameDay(lastReminderEvent.remindedTimestamp, remindedTimestamp)
-            ) {
-                return 0
-            }
-
-            return (lastReminderEvent.processedTimestamp.epochSecond / 60).toInt()
-        }
     }
 }
